+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules/
 /generated/prisma
 .env
+dist/
+tsconfig.tsbuildinfo
