+++ src/api/libs/hashing.ts
@@ -1,6 +1,6 @@
 import bcrypt from "bcrypt";
 
-async function hashing(hashword) {
+async function hashing(hashword: string) {
   try {
     const salt = await bcrypt.genSalt(10);
     const hashedWord = await bcrypt.hash(hashword, salt);
@@ -10,7 +10,7 @@ async function hashing(hashword) {
   }
 }
 
-async function compareWords(word, hashedWord) {
+async function compareWords(word: string, hashedWord: string) {
   return await bcrypt.compare(word, hashedWord);
 }
