+++ src/api/libs/token.ts
@@ -1,7 +1,10 @@
 import jwt from "jsonwebtoken";
 import { ACCESS_TOKEN_SECRET, REFRESH_TOKEN_SECRET } from "./constants.js";
 
-function generateTokens(userId) {
+function generateTokens(userId: number) {
+  if (!ACCESS_TOKEN_SECRET || !REFRESH_TOKEN_SECRET) {
+    throw new Error("토큰 시크릿이 설정되지 않았습니다.");
+  }
   const accessToken = jwt.sign({ id: userId }, ACCESS_TOKEN_SECRET, {
     expiresIn: "1h",
   });
